# app/openapi.yaml
openapi: 3.1.0
info:
  title: Micro-Learning Flashcards API
  description: |
    Flashcard deck management with FSRS spaced repetition scheduling.
    Supports AI-generated system decks and learner-owned personal decks.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /modules/{moduleID}/deck:
    post:
      summary: Generate system deck for a module
      description: Triggers LLM-based card generation from module content via the CMS gRPC service.
      operationId: generateSystemDeck
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/moduleID'
        - name: locale
          in: query
          description: Locale for module content
          schema:
            type: string
            default: en
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Human-readable deck title
      responses:
        '200':
          description: Deck generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckWithCards'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/BadGateway'

    get:
      summary: Get system deck for a module
      operationId: getSystemDeck
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/moduleID'
      responses:
        '200':
          description: System deck with cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckWithCards'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks:
    get:
      summary: List all system decks
      description: Returns all non-deleted system decks with card counts (no card details).
      operationId: listSystemDecks
      tags: [System Decks]
      responses:
        '200':
          description: List of system decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckSummary'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks/{deckID}:
    get:
      summary: Get system deck by ID
      operationId: getSystemDeckByID
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      responses:
        '200':
          description: System deck with cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckWithCards'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a system deck
      description: Sets deleted_at on the deck. User decks that copied it are unaffected.
      operationId: deleteSystemDeck
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      responses:
        '204':
          description: Deck deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks/{deckID}/cards:
    post:
      summary: Add a card to a system deck
      operationId: addCardToSystemDeck
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardInput'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks/{deckID}/cards/{cardID}:
    put:
      summary: Edit a card in a system deck
      description: In-place update. source_concept_id and source_concept_title updated only if non-empty.
      operationId: updateSystemDeckCard
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
        - $ref: '#/components/parameters/cardID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardInput'
      responses:
        '200':
          description: Card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a card from a system deck
      description: Sets deleted_at on the card. User deck junction rows are unaffected — COW cards remain accessible until replaced by an edit.
      operationId: deleteSystemDeckCard
      tags: [System Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
        - $ref: '#/components/parameters/cardID'
      responses:
        '204':
          description: Card deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks/{deckID}/copy:
    post:
      summary: Copy system deck to a personal deck
      description: |
        Creates a learner-owned copy of a system deck using a shallow copy
        (shared card references). Cards are cloned on first edit (copy-on-write).
      operationId: copyDeckToUserDeck
      tags: [Personal Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [learner_id]
              properties:
                learner_id:
                  type: string
                  description: ID of the learner who will own the copy
      responses:
        '201':
          description: Personal deck created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckWithCards'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /decks/{deckID}/review:
    get:
      summary: Get due cards for review
      description: Returns cards that are due for review for the given learner. Includes cards with no SRS record or whose due time has passed.
      operationId: getDueCards
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/deckID'
        - name: learner_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Due cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /learners/{learnerID}/decks:
    get:
      summary: List personal decks for a learner
      operationId: listLearnerDecks
      tags: [Personal Decks]
      parameters:
        - $ref: '#/components/parameters/learnerID'
      responses:
        '200':
          description: Learner's personal decks with cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckWithCards'
        '500':
          $ref: '#/components/responses/InternalError'
  /user-decks/{deckID}:
    delete:
      summary: Soft-delete a personal deck
      description: Sets deleted_at on the deck and any cards it owns.
      operationId: deleteUserDeck
      tags: [Personal Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      responses:
        '204':
          description: Deck deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /user-decks/{deckID}/cards:
    post:
      summary: Add a card to a personal deck
      operationId: addCardToUserDeck
      tags: [Personal Decks]
      parameters:
        - $ref: '#/components/parameters/deckID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCardInput'
      responses:
        '201':
          description: Card added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /cards/{cardID}/review:
    post:
      summary: Submit a review rating for a card
      description: |
        Records a review result and updates the FSRS schedule for the card.
        Rating scale: 1=Again, 2=Hard, 3=Good, 4=Easy.
      operationId: submitCardReview
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/cardID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [learner_id, rating]
              properties:
                learner_id:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 4
                  description: "FSRS rating: 1=Again, 2=Hard, 3=Good, 4=Easy"
      responses:
        '200':
          description: Review recorded; updated SRS state returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SRSCard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    moduleID:
      name: moduleID
      in: path
      required: true
      schema:
        type: string
    deckID:
      name: deckID
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    DeckWithCards:
      allOf:
        - $ref: '#/components/schemas/Deck'
        - type: object
          properties:
            cards:
              type: array
              items:
                $ref: '#/components/schemas/Card'
    DeckSummary:
      allOf:
        - $ref: '#/components/schemas/Deck'
        - type: object
          properties:
            card_count:
              type: integer
              description: Number of cards in the deck
...
    CardInput:
      type: object
      required: [question, correct_answer, distractors, question_ja, correct_answer_ja, distractors_ja]
      properties:
        question:
          type: string
        correct_answer:
          type: string
        distractors:
          type: array
          items:
            type: string
        question_ja:
          type: string
        correct_answer_ja:
          type: string
        distractors_ja:
          type: array
          items:
            type: string
        source_concept_id:
          type: string
        source_concept_title:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: Upstream service error (CMS gRPC)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'